Update README.md
A phishing website is a fraudulent website that is designed to look like a legitimate website with the purpose of tricking individuals into giving away their sensitive information such as login credentials, credit card details, and other personal information. These websites often mimic the design and content of well-known brands or companies in order to gain  the trust of the user.